backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.0.1
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.1
ecdsa==0.19.0
fastapi==0.115.5
h11==0.14.0
idna==3.10
joblib==1.4.2
numpy==2.1.3
pandas==2.2.3
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
xgboost==2.1.2
